FAC-864: Refactor: Change `UseCase` and `UseCaseParamTest` package
