+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/question/GetQuestionIssuesService.java
@@ -5,7 +5,7 @@ import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.domain.Answer;
 import org.flickit.assessment.core.application.domain.ConfidenceLevel;
-import org.flickit.assessment.core.application.port.in.questions.GetQuestionIssuesUseCase;
+import org.flickit.assessment.core.application.port.in.question.GetQuestionIssuesUseCase;
 import org.flickit.assessment.core.application.port.out.answer.LoadAnswerPort;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.flickit.assessment.core.application.port.out.evidence.CountEvidencesPort;
