+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/port/in/question/GetQuestionIssuesUseCaseParamTest.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.port.in.questions;
+package org.flickit.assessment.core.application.port.in.question;
 
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.Test;
