+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/question/GetQuestionIssuesServiceTest.java
@@ -4,7 +4,7 @@ import org.flickit.assessment.common.application.domain.assessment.AssessmentAcc
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.domain.AssessmentResult;
-import org.flickit.assessment.core.application.port.in.questions.GetQuestionIssuesUseCase;
+import org.flickit.assessment.core.application.port.in.question.GetQuestionIssuesUseCase;
 import org.flickit.assessment.core.application.port.out.answer.LoadAnswerPort;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.flickit.assessment.core.application.port.out.evidence.CountEvidencesPort;
